requirements.txt
azure-functions==1.13.3
openai